dashboard/backend/tests/requirements.txt
aprspy
paho-mqtt
python-dotenv
requests
h3
colorlog
sqlalchemy
psycopg2-binary